engine/requirements.txt
pandas
scikit-learn
transformers==4.41.2
torch
numpy
tqdm
flask
gdown
python-dotenv